sdk: tests: refactor: move epics tests to integration
Since they're the only kind of integration tests we have
