+++ raiden-ts/tests/integration/mediate.spec.ts
@@ -9,8 +9,8 @@ import {
   secrethash,
   token,
   tokenNetwork,
-} from '../fixtures';
-import { makeRaidens } from '../mocks';
+} from './fixtures';
+import { makeRaidens } from './mocks';
 
 import { BigNumber } from '@ethersproject/bignumber';
 import { AddressZero, MaxUint256, Zero } from '@ethersproject/constants';
@@ -27,7 +27,7 @@ import { makePaymentId } from '@/transfers/utils';
 import { assert } from '@/utils';
 import { decode, Int, UInt } from '@/utils/types';
 
-import { makeAddress, sleep } from '../../utils';
+import { makeAddress, sleep } from '../utils';
 
 describe('mediate transfers', () => {
   test('success with flat fees', async () => {
