+++ raiden-ts/tests/integration/udc.spec.ts
@@ -1,5 +1,5 @@
-import { confirmationBlocks } from '../fixtures';
-import { makeRaiden, makeRaidens, makeStruct, makeTransaction, waitBlock } from '../mocks';
+import { confirmationBlocks } from './fixtures';
+import { makeRaiden, makeRaidens, makeStruct, makeTransaction, waitBlock } from './mocks';
 
 import { BigNumber } from '@ethersproject/bignumber';
 import { MaxUint256, Zero } from '@ethersproject/constants';
@@ -11,7 +11,7 @@ import { udcDeposit, udcWithdraw, udcWithdrawPlan } from '@/services/actions';
 import { ErrorCodes } from '@/utils/error';
 import type { Hash, UInt } from '@/utils/types';
 
-import { sleep } from '../../utils';
+import { sleep } from '../utils';
 
 test('monitorUdcBalanceEpic', async () => {
   expect.assertions(5);
